LPD-82342 Move tests to objectView file and delete corresponding poshi ones
